requirements.txt
fastapi
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
groq
ragas
langsmith
python-dotenv
pydantic
uvicorn
python-jose
passlib[bcrypt]
python-multipart
bcrypt==4.1.2
